Alterado controller imanges para receber arquivo nulo

+++ app/Http/Controllers/IconController.php
@@ -9,6 +9,7 @@ use App\Models\util\MapError;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Str;
 
 class IconController extends Controller
 {
@@ -20,24 +21,45 @@ class IconController extends Controller
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'descricao' => 'bail|required',
-            'file' => 'bail|required'
+            'descricao' => 'bail|required'
         ],
             [
-                'descricao.required' => 'Descrição é obrigatório!',
-                'file.required' => 'Imagem é obrigatória!'
+                'descricao.required' => 'Descrição é obrigatório!'
             ]);
 
         if ($validator->fails())
             return response()->json(['errors' => MapError::format($validator->messages()), 'status' => 400], 400);
 
-        $icon = Icon::create([
-            'descricao' => $request->descricao,
-            'imagem' => '/icons/' . $request->file->hashName()
-        ]);
+        $mensagem = "Icone criado com sucesso!";
+        $isPresentFile = (isset($request['file']) && $request->hasFile('file'));
 
-        $file = $request->file->store('icons', 'public');
-        return $this->show($icon->id);
+        if (isset($request['id'])) {
+            $iconDB = $this->show($request->id);
+
+            if ($isPresentFile) {
+                if (Storage::disk('public')->exists($iconDB->imagem))
+                    Storage::disk('public')->delete($iconDB->imagem);
+            }
+
+            $mensagem = "Icone atualizado com sucesso!";
+
+            Icon::updateOrCreate(
+                ['id' => isset($request['id']) ? $request->id : null], [
+                'descricao' => isset($request['descricao']) ? $request->descricao : $iconDB->descricao,
+                'imagem' => $isPresentFile ? '/icons/' . $request->file->hashName() : $iconDB->imagem
+            ]);
+        } else {
+            Icon::create([
+                'descricao' => $request->descricao
+            ]);
+        }
+
+        if ($isPresentFile)
+            $file = $request->file->store('icons', 'public');
+
+        return response()->json([
+            'message' => $mensagem,
+            'status' => 200], 200);
     }
 
     public function update(Request $request)
@@ -64,7 +86,9 @@ class IconController extends Controller
             'descricao' => $request->descricao == null ? $icon->descricao : $request->descricao,
             'imagem' => $request->file == null ? $icon->url : '/icons/' . $request->file->hashName()
         ]);
-        return $this->show($icon->id);
+        return response()->json([
+            'message' => 'Icone atualizado com sucesso!',
+            'status' => 200], 200);
     }
 
     public function show($id)
@@ -74,20 +98,29 @@ class IconController extends Controller
 
     public function destroy($id)
     {
+
+        if ($id == 1)
+            return response()->json([
+                'message' => 'Não é possivel remover o icone padrão!',
+                'status' => 400], 400);
+
         $iconDB = Icon::findOrFail($id);
 
+
+        if (!(Str::contains($iconDB->imagem, 'defaul')))
             if (Storage::disk('public')->exists($iconDB->imagem))
                 Storage::disk('public')->delete($iconDB->imagem);
 
-            $iconDB->delete();
+        $iconDB->delete();
 
-            return response()->json([
-                'message' => 'Icon removida com sucesso!',
-                'status' => 200], 200);
+        return response()->json([
+            'message' => 'Icone removido com sucesso!',
+            'status' => 200], 200);
     }
 
 
-    public function associar(Request $request){
+    public function associar(Request $request)
+    {
 
         Catalogo::findOrFail($request->catalogo_id)
             ->icon()
